@@ -18,13 +18,19 @@
         }
     ],
     "require": {
-        "php": ">=8.0",
-        "illuminate/support": "^8.0|^9.0|^10.0",
-        "illuminate/cache": "^8.0|^9.0|^10.0",
-        "intervention/image": "^2.7"
+        "php": ">=8.1",
+        "illuminate/support": "^10.0|^11.0",
+        "illuminate/cache": "^10.0|^11.0",
+        "intervention/image": "^3.4"
     },
+    "suggest": {
+        "ext-gd": "Neede to support image manipulation",
+        "ext-imagick": "Neede to support image manipulation, better than GD"
+    },
+
     "require-dev": {
-        "phpunit/phpunit": "~9.0",
+        "roave/security-advisories": "dev-latest",
+        "phpunit/phpunit": "^10.5",
         "mockery/mockery": "~1.3",
         "php-coveralls/php-coveralls": "^2.1"
     },
